fix: rename hooks
